+++ src/glyphtune/io/stream.py
@@ -1,22 +1,23 @@
 """Streaming audio output."""
 
+import sys
 import numpy as np
 import pyaudio
-from glyphtune import waveforms
+from glyphtune import signal, waveforms
 
 
-class Stream:
-    """Audio output stream."""
+class IOStream:
+    """Audio I/O stream."""
 
     def __init__(
         self, sampling_rate: int = 44100, buffer_size: int = 512, channels: int = 2
     ) -> None:
-        """Initializes an audio stream.
+        """Initializes an audio I/O stream.
 
         Args:
             sampling_rate: the sampling rate to use in samples per second.
             buffer_size: the size of chunks to be streamed in samples.
-            channels: number of channels to output.
+            channels: number of channels.
         """
         self.sampling_rate = sampling_rate
         self.buffer_size = buffer_size
@@ -24,7 +25,7 @@ class Stream:
 
     @property
     def sampling_rate(self) -> int:
-        """Sampling rate of the output stream."""
+        """Sampling rate of the stream."""
         return self.__sampling_rate
 
     @sampling_rate.setter
@@ -35,7 +36,7 @@ class Stream:
 
     @property
     def buffer_size(self) -> int:
-        """Buffer size of the output stream."""
+        """Buffer size of the stream."""
         return self.__buffer_size
 
     @buffer_size.setter
@@ -46,7 +47,7 @@ class Stream:
 
     @property
     def channels(self) -> int:
-        """Number of channels of the output stream."""
+        """Number of channels of the stream."""
         return self.__channels
 
     @channels.setter
@@ -55,8 +56,8 @@ class Stream:
             raise ValueError("Number of channels must be positive")
         self.__channels = value
 
-    def stream_waveform(self, waveform: waveforms.Waveform) -> None:
-        """Stream a waveform until `SystemExit` or `KeyboardInterrupt` is raised.
+    def play(self, waveform: waveforms.Waveform) -> None:
+        """Samples and streams a waveform's output until interrupted.
 
         Args:
             waveform: waveform to be streamed.
@@ -81,3 +82,27 @@ class Stream:
             stream.stop_stream()
             stream.close()
             py_audio.terminate()
+
+    def record(self) -> signal.Signal:
+        """Returns signal of audio input that is recorded until interrupted."""
+        py_audio = pyaudio.PyAudio()
+        stream = py_audio.open(
+            self.sampling_rate, self.channels, pyaudio.paFloat32, input=True
+        )
+        read_bytes = bytes()
+        chunks = 0
+        try:
+            while True:
+                read_bytes += stream.read(self.buffer_size)
+                chunks += 1
+        except (KeyboardInterrupt, SystemExit) as exception:
+            stream.stop_stream()
+            stream.close()
+            py_audio.terminate()
+            if isinstance(exception, SystemExit):
+                sys.exit()
+        read_array_flat = np.frombuffer(read_bytes, dtype=np.float32)
+        read_signal = signal.Signal(
+            read_array_flat.reshape((self.channels, self.buffer_size * chunks))
+        )
+        return read_signal