+++ CHANGELOG.md
@@ -12,8 +12,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Effects module (including base `Effect` class).
 - Stereo effects module (including `StereoPan`, `StereoLevels`, `StereoInterMix`, and `StereoDelay`).
 - Resampling module (including `ResampleWaveform` class).
-- Output stream with an arbitrary number of channels (`glyphtune.output.Stream`).
-- Output to wav file (`glyphtune.output.write_wav` and `glyphtune.output.WavParameters`)
+- Audio stream I/O with an arbitrary number of channels (`glyphtune.io.IOStream`).
+- wav file I/O (`glyphtune.io.read_wav`, `glyphtune.io.write_wav` and `glyphtune.io.WavParameters`)
 - `glyphtune.signal` module (including `Signal` class).
 
 ### Changed
@@ -21,6 +21,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `Waveform.sample_arr` is renamed to `sample_time` and is now expected to be able to take as an argument a `Signal` object of the shape `(channels, samples)` and return a `Signal` object of the same shape.
 - `Waveform.sample_seconds` and `Waveform.sample_samples` now accept a `channels` option and can sample multi-channel audio, returning `Signal` objects of the shape `(channels, samples)`.
 - `Waveform.sample_seconds` and `Waveform.sample_samples` now sample stereo audio by default.
+- `glyphtune.output` module moved to `glyphtune.io`.
 - `PhaseModulation` no longer accepts the `frequency_modulation` option and is no longer able to perform frequency modulation.
 - `PeriodicWave` now properly overrides `__eq__` and `__repr__`.
 - Periodic wave representations now explicitly state `phase` keyword.
@@ -29,7 +30,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Removed
 
-- `glyphtune.output.MonoStream` class (replaced by `glyphtune.output.Stream`).
+- `glyphtune.output.MonoStream` class (replaced by `glyphtune.io.IOStream`).
 - Calculus module (`glyphtune.waveforms.calculus`, including `DerivativeWaveform` and `IntegralWaveform`).
 - Frequency modulation (`glyphtune.waveforms.modulation.frequency_modulate`).
 - `glyphtune.FloatArray` type alias.